refactor: unify realtime hubs
